Initial commit
Add Go code for file shredder

This very simple program executes a superficial shred of a file passed
on the command line.

This kind of shred exemplifies basic programming language routines but
is far from guaranteed to perform an actual real data shredding.
Other layers of the OS and the hardware are designed to guarantee data
survival and as such will not really allow for true data shredding.